Move storage import/export enums to contracts
