+++ Veriado.Application/Abstractions/ImportModels.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Veriado.Contracts.Storage;
 
 namespace Veriado.Application.Abstractions;
 
@@ -19,38 +20,6 @@ public sealed record ImportRequest
         = null;
 }
 
-public enum ImportConflictStrategy
-{
-    SkipIfExists,
-    UpdateIfNewer,
-    AlwaysOverwrite,
-    CreateDuplicate,
-}
-
-public enum ImportIssueType
-{
-    PackageMissing,
-    ManifestMissing,
-    ManifestUnsupported,
-    MetadataMissing,
-    MetadataUnsupported,
-    FilesRootMissing,
-    MissingDescriptor,
-    MissingFile,
-    SizeMismatch,
-    HashMismatch,
-    SchemaUnsupported,
-    FileCountMismatch,
-    FileBytesMismatch,
-    ConflictExistingFile,
-}
-
-public enum ImportIssueSeverity
-{
-    Warning,
-    Error,
-}
-
 public sealed record ImportValidationIssue(
     ImportIssueType Type,
     ImportIssueSeverity Severity,
@@ -66,15 +35,6 @@ public sealed record ValidatedImportFile(
     long SizeBytes,
     string? MimeType,
     DateTimeOffset LastModifiedAtUtc);
-
-public enum ImportItemStatus
-{
-    New,
-    DuplicateSameVersion,
-    DuplicateOlderInDb,
-    DuplicateNewerInDb,
-    ConflictOther,
-}
 
 public sealed record ImportItemPreview(
     Guid FileId,
@@ -110,13 +70,6 @@ public sealed record ImportValidationResult(
             0,
             0,
             0);
-}
-
-public enum ImportCommitStatus
-{
-    Success,
-    PartialSuccess,
-    Failed,
 }
 
 public sealed record ImportCommitResult(
